requirements.txt
google-genai>=1.0.0
requests>=2.32.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
tenacity>=9.0.0
